common_meta
Module datanode

Structs

DATANODE_LEASE_KEY_PATTERN 🔒

DATANODE_STAT_KEY_PATTERN 🔒

DatanodeStatKey

The key of the datanode stat in the memory store.

DatanodeStatValue

The value of the datanode stat in the memory store.

INACTIVE_REGION_KEY_PATTERN 🔒

RegionStat

The statistics of a region.

Stat

The key of the datanode stat in the storage.

Constants

DATANODE_LEASE_PREFIX 🔒

DATANODE_STAT_PREFIX 🔒

INACTIVE_REGION_PREFIX 🔒

REGION_STATISTIC_KEY