common_meta::ddl

Trait RegionFailureDetectorController

pub trait RegionFailureDetectorController: Send + Sync {
    // Required methods
    fn register_failure_detectors<'life0, 'async_trait>(
        &'life0 self,
        detecting_regions: Vec<DetectingRegion>,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn deregister_failure_detectors<'life0, 'async_trait>(
        &'life0 self,
        detecting_regions: Vec<DetectingRegion>,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Used for actively registering Region failure detectors.

Ensuring the Region Supervisor can detect Region failures without relying on the first heartbeat from the datanode.

Required Methods

fn register_failure_detectors<'life0, 'async_trait>( &'life0 self, detecting_regions: Vec<DetectingRegion>, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Registers failure detectors for the given identifiers.

fn deregister_failure_detectors<'life0, 'async_trait>( &'life0 self, detecting_regions: Vec<DetectingRegion>, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Deregisters failure detectors for the given identifiers.