common_meta::ddl::utils::table_info

Function batch_update_table_info_values

pub(crate) async fn batch_update_table_info_values(
    table_metadata_manager: &TableMetadataManager,
    table_info_values: Vec<(DeserializedValueWithBytes<TableInfoValue>, RawTableInfo)>,
) -> Result<()>

Batch updates the table info values.

The table info values are grouped into chunks, and each chunk is updated in a single transaction.

Returns an error if any table info value fails to update.