common_meta::election

Trait LeaderKey

pub trait LeaderKey:
    Send
    + Sync
    + Debug {
    // Required methods
    fn name(&self) -> &[u8] ;
    fn key(&self) -> &[u8] ;
    fn revision(&self) -> i64;
    fn lease_id(&self) -> i64;
}

LeaderKey is a key that represents the leader of metasrv. The structure is corresponding to [etcd_client::LeaderKey].

Required Methods

fn name(&self) -> &[u8]

The name in byte. name is the election identifier that corresponds to the leadership key.

fn key(&self) -> &[u8]

The key in byte. key is an opaque key representing the ownership of the election. If the key is deleted, then leadership is lost.

fn revision(&self) -> i64

The creation revision of the key.

fn lease_id(&self) -> i64

The lease ID of the election leader.