common_meta::error

Enum Error

pub enum Error {

    EmptyKey {
        location: Location,
    },
    RegionOperatingRace {
        location: Location,
        peer_id: DatanodeId,
        region_id: RegionId,
    },
    ConnectEtcd {
        error: Error,
        location: Location,
    },
    EtcdFailed {
        error: Error,
        location: Location,
    },
    EtcdTxnFailed {
        max_operations: usize,
        error: Error,
        location: Location,
    },
    NextSequence {
        err_msg: String,
        location: Location,
    },
    UnexpectedSequenceValue {
        err_msg: String,
        location: Location,
    },
    TableInfoNotFound {
        table: String,
        location: Location,
    },
    RegisterProcedureLoader {
        type_name: String,
        location: Location,
        source: Error,
    },
    RegisterRepartitionProcedureLoader {
        location: Location,
        source: BoxedError,
    },
    CreateRepartitionProcedure {
        source: BoxedError,
        location: Location,
    },
    SubmitProcedure {
        location: Location,
        source: Error,
    },
    QueryProcedure {
        location: Location,
        source: Error,
    },
    ProcedureNotFound {
        location: Location,
        pid: String,
    },
    ParseProcedureId {
        location: Location,
        key: String,
        error: ParseIdError,
    },
    Unsupported {
        operation: String,
        location: Location,
    },
    ProcedureStateReceiver {
        procedure_id: ProcedureId,
        location: Location,
        source: Error,
    },
    ProcedureStateReceiverNotFound {
        procedure_id: ProcedureId,
        location: Location,
    },
    WaitProcedure {
        location: Location,
        source: Error,
    },
    StartProcedureManager {
        location: Location,
        source: Error,
    },
    StopProcedureManager {
        location: Location,
        source: Error,
    },
    ProcedureOutput {
        procedure_id: String,
        err_msg: String,
        location: Location,
    },
    PrimaryKeyNotFound {
        key: String,
        location: Location,
    },
    BuildTableMeta {
        table_name: String,
        error: TableMetaBuilderError,
        location: Location,
    },
    Table {
        location: Location,
        source: Error,
    },
    TableRouteNotFound {
        table_id: TableId,
        location: Location,
    },
    TableRepartNotFound {
        table_id: TableId,
        location: Location,
    },
    DecodeProto {
        location: Location,
        error: DecodeError,
    },
    EncodeJson {
        location: Location,
        error: Error,
    },
    DecodeJson {
        location: Location,
        error: Error,
    },
    SerializeToJson {
        input: String,
        error: Error,
        location: Location,
    },
    DeserializeFromJson {
        input: String,
        error: Error,
        location: Location,
    },
    PayloadNotExist {
        location: Location,
    },
    SerdeJson {
        error: Error,
        location: Location,
    },
    ParseOption {
        key: String,
        value: String,
        location: Location,
    },
    RouteInfoCorrupted {
        err_msg: String,
        location: Location,
    },
    IllegalServerState {
        code: i32,
        err_msg: String,
        location: Location,
    },
    ConvertAlterTableRequest {
        source: Error,
        location: Location,
    },
    InvalidProtoMsg {
        err_msg: String,
        location: Location,
    },
    Unexpected {
        err_msg: String,
        location: Location,
    },
    TableAlreadyExists {
        table_name: String,
        location: Location,
    },
    ViewAlreadyExists {
        view_name: String,
        location: Location,
    },
    FlowAlreadyExists {
        flow_name: String,
        location: Location,
    },
    SchemaAlreadyExists {
        catalog: String,
        schema: String,
        location: Location,
    },
    ConvertRawKey {
        location: Location,
        error: Utf8Error,
    },
    TableNotFound {
        table_name: String,
        location: Location,
    },
    RegionNotFound {
        region_id: RegionId,
        location: Location,
    },
    ViewNotFound {
        view_name: String,
        location: Location,
    },
    FlowNotFound {
        flow_name: String,
        location: Location,
    },
    FlowRouteNotFound {
        flow_name: String,
        location: Location,
    },
    SchemaNotFound {
        table_schema: String,
        location: Location,
    },
    CatalogNotFound {
        catalog: String,
        location: Location,
    },
    InvalidMetadata {
        err_msg: String,
        location: Location,
    },
    InvalidViewInfo {
        err_msg: String,
        location: Location,
    },
    InvalidFlowRequestBody {
        body: Box<Option<Body>>,
        location: Location,
    },
    GetKvCache {
        err_msg: String,
    },
    CacheNotGet {
        key: String,
        location: Location,
    },
    EtcdTxnOpResponse {
        err_msg: String,
        location: Location,
    },
    External {
        location: Location,
        source: BoxedError,
    },
    ResponseExceededSizeLimit {
        location: Location,
        source: BoxedError,
    },
    InvalidHeartbeatResponse {
        location: Location,
    },
    OperateDatanode {
        location: Location,
        peer: Peer,
        source: BoxedError,
    },
    RetryLater {
        source: BoxedError,
        clean_poisons: bool,
    },
    AbortProcedure {
        location: Location,
        source: BoxedError,
        clean_poisons: bool,
    },
    EncodeWalOptions {
        wal_options: WalOptions,
        error: Error,
        location: Location,
    },
    InvalidNumTopics {
        num_topics: usize,
        location: Location,
    },
    BuildKafkaClient {
        broker_endpoints: Vec<String>,
        location: Location,
        error: Error,
    },
    TlsConfig {
        location: Location,
        source: Error,
    },
    BuildKafkaCtrlClient {
        location: Location,
        error: Error,
    },
    KafkaPartitionClient {
        topic: String,
        partition: i32,
        location: Location,
        error: Error,
    },
    KafkaGetOffset {
        topic: String,
        partition: i32,
        location: Location,
        error: Error,
    },
    ProduceRecord {
        topic: String,
        location: Location,
        error: Error,
    },
    CreateKafkaWalTopic {
        location: Location,
        error: Error,
    },
    EmptyTopicPool {
        location: Location,
    },
    UnexpectedLogicalRouteTable {
        location: Location,
        err_msg: String,
    },
    EmptyDdlTasks {
        name: String,
        location: Location,
    },
    MetadataCorruption {
        err_msg: String,
        location: Location,
    },
    AlterLogicalTablesInvalidArguments {
        err_msg: String,
        location: Location,
    },
    CreateLogicalTablesInvalidArguments {
        err_msg: String,
        location: Location,
    },
    InvalidNodeInfoKey {
        key: String,
        location: Location,
    },
    InvalidStatKey {
        key: String,
        location: Location,
    },
    ParseNum {
        err_msg: String,
        error: ParseIntError,
        location: Location,
    },
    InvalidRole {
        role: i32,
        location: Location,
    },
    InvalidSetDatabaseOption {
        key: String,
        value: String,
        location: Location,
    },
    InvalidUnsetDatabaseOption {
        key: String,
        location: Location,
    },
    MismatchPrefix {
        prefix: String,
        key: String,
        location: Location,
    },
    MoveValues {
        err_msg: String,
        location: Location,
    },
    FromUtf8 {
        name: String,
        error: FromUtf8Error,
        location: Location,
    },
    ValueNotExist {
        location: Location,
    },
    GetCache {
        source: Arc<Error>,
    },
    PostgresExecution {
        sql: String,
        error: Error,
        location: Location,
    },
    CreatePostgresPool {
        error: CreatePoolError,
        location: Location,
    },
    GetPostgresConnection {
        reason: String,
        location: Location,
    },
    PostgresTransaction {
        error: Error,
        location: Location,
        operation: String,
    },
    PostgresTlsConfig {
        reason: String,
        location: Location,
    },
    LoadTlsCertificate {
        path: String,
        error: Error,
        location: Location,
    },
    InvalidTlsConfig {
        reason: String,
        location: Location,
    },
    MySqlExecution {
        sql: String,
        error: Error,
        location: Location,
    },
    CreateMySqlPool {
        error: Error,
        location: Location,
    },
    MySqlTransaction {
        error: Error,
        location: Location,
        operation: String,
    },
    RdsTransactionRetryFailed {
        location: Location,
    },
    DatanodeTableInfoNotFound {
        datanode_id: DatanodeId,
        table_id: TableId,
        location: Location,
    },
    InvalidTopicNamePrefix {
        prefix: String,
        location: Location,
    },
    ParseWalOptions {
        wal_options: String,
        location: Location,
        error: Error,
    },
    NoLeader {
        table_id: TableId,
        location: Location,
    },
    ProcedurePoisonConflict {
        key: String,
        value: String,
        location: Location,
    },
    PutPoison {
        location: Location,
        source: Error,
    },
    InvalidFilePath {
        location: Location,
        file_path: String,
    },
    SerializeFlexbuffers {
        location: Location,
        error: SerializationError,
    },
    DeserializeFlexbuffers {
        location: Location,
        error: DeserializationError,
    },
    ReadFlexbuffers {
        location: Location,
        error: ReaderError,
    },
    InvalidFileName {
        location: Location,
        reason: String,
    },
    InvalidFileExtension {
        location: Location,
        reason: String,
    },
    WriteObject {
        location: Location,
        file_path: String,
        error: Error,
    },
    ReadObject {
        location: Location,
        file_path: String,
        error: Error,
    },
    MissingColumnIds {
        location: Location,
    },
    MissingColumnInColumnMetadata {
        column_name: String,
        location: Location,
        table_name: String,
        table_id: TableId,
    },
    MismatchColumnId {
        column_name: String,
        column_id: u32,
        location: Location,
        table_name: String,
        table_id: TableId,
    },
    ConvertColumnDef {
        column: String,
        location: Location,
        source: Error,
    },
    ConvertTimeRanges {
        location: Location,
        source: Error,
    },
    ColumnMetadataConflicts {
        table_name: String,
        table_id: TableId,
    },
    ColumnNotFound {
        column_name: String,
        column_id: u32,
    },
    ColumnIdMismatch {
        column_name: String,
        expected_column_id: u32,
        actual_column_id: u32,
    },
    TimestampMismatch {
        expected_column_name: String,
        expected_column_id: u32,
        actual_column_name: String,
        actual_column_id: u32,
    },

}

Variants

EmptyKey

Fields

location: Location

RegionOperatingRace

Fields

location: Location

peer_id: DatanodeId

region_id: RegionId

ConnectEtcd

Fields

error: Error

location: Location

EtcdFailed

Fields

error: Error

location: Location

EtcdTxnFailed

Fields

max_operations: usize

error: Error

location: Location

NextSequence

Fields

err_msg: String

location: Location

UnexpectedSequenceValue

Fields

err_msg: String

location: Location

TableInfoNotFound

Fields

table: String

location: Location

RegisterProcedureLoader

Fields

type_name: String

location: Location

source: Error

RegisterRepartitionProcedureLoader

Fields

location: Location

source: BoxedError

CreateRepartitionProcedure

Fields

source: BoxedError

location: Location

SubmitProcedure

Fields

location: Location

source: Error

QueryProcedure

Fields

location: Location

source: Error

ProcedureNotFound

Fields

location: Location

pid: String

ParseProcedureId

Fields

location: Location

key: String

error: ParseIdError

Unsupported

Fields

operation: String

location: Location

ProcedureStateReceiver

Fields

procedure_id: ProcedureId

location: Location

source: Error

ProcedureStateReceiverNotFound

Fields

procedure_id: ProcedureId

location: Location

WaitProcedure

Fields

location: Location

source: Error

StartProcedureManager

Fields

location: Location

source: Error

StopProcedureManager

Fields

location: Location

source: Error

ProcedureOutput

Fields

procedure_id: String

err_msg: String

location: Location

PrimaryKeyNotFound

Fields

key: String

location: Location

BuildTableMeta

Fields

table_name: String

error: TableMetaBuilderError

location: Location

Table

Fields

location: Location

source: Error

TableRouteNotFound

Fields

table_id: TableId

location: Location

TableRepartNotFound

Fields

table_id: TableId

location: Location

DecodeProto

Fields

location: Location

error: DecodeError

EncodeJson

Fields

location: Location

error: Error

DecodeJson

Fields

location: Location

error: Error

SerializeToJson

Fields

input: String

error: Error

location: Location

DeserializeFromJson

Fields

input: String

error: Error

location: Location

PayloadNotExist

Fields

location: Location

SerdeJson

Fields

error: Error

location: Location

ParseOption

Fields

key: String

value: String

location: Location

RouteInfoCorrupted

Fields

err_msg: String

location: Location

IllegalServerState

Fields

code: i32

err_msg: String

location: Location

ConvertAlterTableRequest

Fields

source: Error

location: Location

InvalidProtoMsg

Fields

err_msg: String

location: Location

Unexpected

Fields

err_msg: String

location: Location

TableAlreadyExists

Fields

table_name: String

location: Location

ViewAlreadyExists

Fields

view_name: String

location: Location

FlowAlreadyExists

Fields

flow_name: String

location: Location

SchemaAlreadyExists

Fields

catalog: String

schema: String

location: Location

ConvertRawKey

Fields

location: Location

error: Utf8Error

TableNotFound

Fields

table_name: String

location: Location

RegionNotFound

Fields

region_id: RegionId

location: Location

ViewNotFound

Fields

view_name: String

location: Location

FlowNotFound

Fields

flow_name: String

location: Location

FlowRouteNotFound

Fields

flow_name: String

location: Location

SchemaNotFound

Fields

table_schema: String

location: Location

CatalogNotFound

Fields

catalog: String

location: Location

InvalidMetadata

Fields

err_msg: String

location: Location

InvalidViewInfo

Fields

err_msg: String

location: Location

InvalidFlowRequestBody

Fields

body: Box<Option<Body>>

location: Location

GetKvCache

Fields

err_msg: String

CacheNotGet

Fields

key: String

location: Location

EtcdTxnOpResponse

Fields

err_msg: String

location: Location

External

Fields

location: Location

source: BoxedError

ResponseExceededSizeLimit

Fields

location: Location

source: BoxedError

InvalidHeartbeatResponse

Fields

location: Location

OperateDatanode

Fields

location: Location

peer: Peer

source: BoxedError

RetryLater

Fields

source: BoxedError

clean_poisons: bool

AbortProcedure

Fields

location: Location

source: BoxedError

clean_poisons: bool

EncodeWalOptions

Fields

wal_options: WalOptions

error: Error

location: Location

InvalidNumTopics

Fields

num_topics: usize

location: Location

BuildKafkaClient

Fields

broker_endpoints: Vec<String>

location: Location

error: Error

TlsConfig

Fields

location: Location

source: Error

BuildKafkaCtrlClient

Fields

location: Location

error: Error

KafkaPartitionClient

Fields

topic: String

partition: i32

location: Location

error: Error

KafkaGetOffset

Fields

topic: String

partition: i32

location: Location

error: Error

ProduceRecord

Fields

topic: String

location: Location

error: Error

CreateKafkaWalTopic

Fields

location: Location

error: Error

EmptyTopicPool

Fields

location: Location

UnexpectedLogicalRouteTable

Fields

location: Location

err_msg: String

EmptyDdlTasks

Fields

name: String

location: Location

MetadataCorruption

Fields

err_msg: String

location: Location

AlterLogicalTablesInvalidArguments

Fields

err_msg: String

location: Location

CreateLogicalTablesInvalidArguments

Fields

err_msg: String

location: Location

InvalidNodeInfoKey

Fields

key: String

location: Location

InvalidStatKey

Fields

key: String

location: Location

ParseNum

Fields

err_msg: String

error: ParseIntError

location: Location

InvalidRole

Fields

role: i32

location: Location

InvalidSetDatabaseOption

Fields

key: String

value: String

location: Location

InvalidUnsetDatabaseOption

Fields

key: String

location: Location

MismatchPrefix

Fields

prefix: String

key: String

location: Location

MoveValues

Fields

err_msg: String

location: Location

FromUtf8

Fields

name: String

error: FromUtf8Error

location: Location

ValueNotExist

Fields

location: Location

GetCache

Fields

source: Arc<Error>

PostgresExecution

Fields

sql: String

error: Error

location: Location

CreatePostgresPool

Fields

error: CreatePoolError

location: Location

GetPostgresConnection

Fields

reason: String

location: Location

PostgresTransaction

Fields

error: Error

location: Location

operation: String

PostgresTlsConfig

Fields

reason: String

location: Location

LoadTlsCertificate

Fields

path: String

error: Error

location: Location

InvalidTlsConfig

Fields

reason: String

location: Location

MySqlExecution

Fields

sql: String

error: Error

location: Location

CreateMySqlPool

Fields

error: Error

location: Location

MySqlTransaction

Fields

error: Error

location: Location

operation: String

RdsTransactionRetryFailed

Fields

location: Location

DatanodeTableInfoNotFound

Fields

datanode_id: DatanodeId

table_id: TableId

location: Location

InvalidTopicNamePrefix

Fields

prefix: String

location: Location

ParseWalOptions

Fields

wal_options: String

location: Location

error: Error

NoLeader

Fields

table_id: TableId

location: Location

ProcedurePoisonConflict

Fields

key: String

value: String

location: Location

PutPoison

Fields

location: Location

source: Error

InvalidFilePath

Fields

location: Location

file_path: String

SerializeFlexbuffers

Fields

location: Location

error: SerializationError

DeserializeFlexbuffers

Fields

location: Location

error: DeserializationError

ReadFlexbuffers

Fields

location: Location

error: ReaderError

InvalidFileName

Fields

location: Location

reason: String

InvalidFileExtension

Fields

location: Location

reason: String

WriteObject

Fields

location: Location

file_path: String

error: Error

ReadObject

Fields

location: Location

file_path: String

error: Error

MissingColumnIds

Fields

location: Location

MissingColumnInColumnMetadata

Fields

column_name: String

location: Location

table_name: String

table_id: TableId

MismatchColumnId

Fields

column_name: String

column_id: u32

location: Location

table_name: String

table_id: TableId

ConvertColumnDef

Fields

column: String

location: Location

source: Error

ConvertTimeRanges

Fields

location: Location

source: Error

ColumnMetadataConflicts

Fields

table_name: String

table_id: TableId

ColumnNotFound

Fields

column_name: String

column_id: u32

ColumnIdMismatch

Fields

column_name: String

expected_column_id: u32

actual_column_id: u32

TimestampMismatch

Fields

expected_column_name: String

expected_column_id: u32

actual_column_name: String

actual_column_id: u32

Implementations

impl Error

pub fn is_serialization_error(&self) -> bool

Check if the error is a serialization error.

pub fn retry_later<E: ErrorExt + Send + Sync + 'static>(err: E) -> Error

Creates a new Error::RetryLater error from source err.

pub fn is_retry_later(&self) -> bool

Determine whether it is a retry later type through [StatusCode]

pub fn need_clean_poisons(&self) -> bool

Determine whether it needs to clean poisons.

pub fn is_exceeded_size_limit(&self) -> bool

Returns true if the response exceeds the size limit.