Module common_meta::key

This mod defines all the keys used in the metadata store (Metasrv). Specifically, there are these kinds of keys:

Datanode table key: __dn_table/{datanode_id}/{table_id}

The value is a DatanodeTableValue struct; it contains table_id and the regions that belong to this Datanode.

This key is primary used in the startup of Datanode, to let Datanode know which tables and regions it should open.

Table info key: __table_info/{table_id}

The value is a TableInfoValue struct; it contains the whole table info (like column schemas).

This key is mainly used in constructing the table in Datanode and Frontend.

Catalog name key: __catalog_name/{catalog_name}

Indices all catalog names

Schema name key: __schema_name/{catalog_name}/{schema_name}

Indices all schema names belong to the {catalog_name}

Table name key: __table_name/{catalog_name}/{schema_name}/{table_name}

The value is a TableNameValue struct; it contains the table id.

Used in the table name to table id lookup.

All keys have related managers. The managers take care of the serialization and deserialization of keys and values, and the interaction with the underlying KV store backend.

To simplify the managers used in struct fields and function parameters, we define a “unify” table metadata manager: TableMetadataManager. It contains all the managers defined above. It’s recommended to just use this manager only.

Modules

catalog_name

datanode_table

schema_name

table_info

table_name

table_region

table_route

test_utils

txn_helper 🔒

Macros

ensure_values 🔒

Structs

CATALOG_NAME_KEY_PATTERN 🔒

CATALOG_NAME_KEY: {CATALOG_NAME_KEY_PREFIX}/{catalog_name}

DATANODE_TABLE_KEY_PATTERN 🔒

DeserializedValueWithBytes

A struct containing a deserialized value(inner) and an original bytes.

SCHEMA_NAME_KEY_PATTERN 🔒

SCHEMA_NAME_KEY: {SCHEMA_NAME_KEY_PREFIX}/{catalog_name}/{schema_name}

TABLE_NAME_KEY_PATTERN 🔒

TableMetadataManager

Constants

CACHE_KEY_PREFIXES

CATALOG_NAME_KEY_PREFIX

DATANODE_TABLE_KEY_PREFIX 🔒

MAINTENANCE_KEY

NAME_PATTERN

REMOVED_PREFIX

SCHEMA_NAME_KEY_PREFIX

TABLE_INFO_KEY_PREFIX

TABLE_NAME_KEY_PREFIX

TABLE_REGION_KEY_PREFIX 🔒

TABLE_ROUTE_PREFIX

Traits

TableMetaKey

TableMetaValue

Type Aliases

RegionDistribution

TableMetadataManagerRef