common_meta::key

Struct DeserializedValueWithBytes

pub struct DeserializedValueWithBytes<T: DeserializeOwned + Serialize> {
    bytes: Bytes,
    inner: T,
}

A struct containing a deserialized value(inner) and an original bytes.

Serialize behaviors:

The inner field will be ignored.

Deserialize behaviors:

The inner field will be deserialized from the bytes field.

Fields
bytes: Bytesinner: T
Implementations

impl<T: Serialize + DeserializeOwned + MetadataValue> DeserializedValueWithBytes<T>

pub fn from_inner_bytes(bytes: Bytes) -> Result<Self>

Returns a struct containing a deserialized value and an original bytes. It accepts original bytes of inner.

pub fn from_inner_slice(bytes: &[u8]) -> Result<Self>

Returns a struct containing a deserialized value and an original bytes. It accepts original bytes of inner.

pub fn into_inner(self) -> T

pub fn get_inner_ref(&self) -> &T

pub fn get_raw_bytes(&self) -> Vec<u8>

Returns original bytes

pub fn from_inner(inner: T) -> Self