common_meta::key

Module table_repart

Structs

TableRepartKey

The key stores table repartition metadata. Specifically, it records the relation between source and destination regions after a repartition operation is completed. This is distinct from the initial partitioning scheme of the table. For example, after repartition, a destination region may still hold files from a source region; this mapping should be updated once repartition is done. The GC scheduler uses this information to clean up those files (and removes this mapping if all files from the source region are cleaned).

TableRepartManager

TableRepartValue

Functions

table_repart_decoder

Decodes KeyValue to TableRepartKey and TableRepartValue.

Type Aliases

TableRepartValueDecodeResult