common_meta::kv_backend

Module rds

Modules

mysql 🔒

postgres 🔒

Structs

RdsStore

Rds backed store for metsrv

Enums

ExecutorImpl

Constants

RDS_STORE_OP_BATCH_DELETE 🔒

RDS_STORE_OP_BATCH_GET 🔒

RDS_STORE_OP_BATCH_PUT 🔒

RDS_STORE_OP_RANGE_DELETE 🔒

RDS_STORE_OP_RANGE_QUERY 🔒

RDS_STORE_TXN_RETRY_COUNT 🔒

Traits

Executor

Query executor for rds. It can execute queries or generate a transaction executor.

ExecutorFactory

Factory for creating default and transaction query executors.

KvQueryExecutor

Transaction

Transaction query executor for rds. It can execute queries in transaction or commit the transaction.

Functions

check_txn_ops 🔒

Checks if the transaction operations are the same type.

Type Aliases

MySqlStore

A MySQL-backed key-value store. It uses [sqlx::Pool] as the connection pool for RdsStore.

PgStore

A PostgreSQL-backed key-value store for metasrv. It uses [deadpool_postgres::Pool] as the connection pool for RdsStore.