Trait common_meta::kv_backend::KvBackend

pub trait KvBackend: TxnServicewhere
    Self::Error: ErrorExt,{

    // Required methods
    fn name(&self) -> &str;
    fn as_any(&self) -> &dyn Any;
    fn range<'life0, 'async_trait>(
        &'life0 self,
        req: RangeRequest
    ) -> Pin<Box<dyn Future<Output = Result<RangeResponse, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn put<'life0, 'async_trait>(
        &'life0 self,
        req: PutRequest
    ) -> Pin<Box<dyn Future<Output = Result<PutResponse, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn batch_put<'life0, 'async_trait>(
        &'life0 self,
        req: BatchPutRequest
    ) -> Pin<Box<dyn Future<Output = Result<BatchPutResponse, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn batch_get<'life0, 'async_trait>(
        &'life0 self,
        req: BatchGetRequest
    ) -> Pin<Box<dyn Future<Output = Result<BatchGetResponse, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn compare_and_put<'life0, 'async_trait>(
        &'life0 self,
        req: CompareAndPutRequest
    ) -> Pin<Box<dyn Future<Output = Result<CompareAndPutResponse, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn delete_range<'life0, 'async_trait>(
        &'life0 self,
        req: DeleteRangeRequest
    ) -> Pin<Box<dyn Future<Output = Result<DeleteRangeResponse, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn batch_delete<'life0, 'async_trait>(
        &'life0 self,
        req: BatchDeleteRequest
    ) -> Pin<Box<dyn Future<Output = Result<BatchDeleteResponse, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn move_value<'life0, 'async_trait>(
        &'life0 self,
        req: MoveValueRequest
    ) -> Pin<Box<dyn Future<Output = Result<MoveValueResponse, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn get<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Option<KeyValue>, Self::Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn exists<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn delete<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 [u8],
        prev_kv: bool
    ) -> Pin<Box<dyn Future<Output = Result<Option<KeyValue>, Self::Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }

}

Required Methods

fn name(&self) -> &str

fn as_any(&self) -> &dyn Any

fn range<'life0, 'async_trait>( &'life0 self, req: RangeRequest ) -> Pin<Box<dyn Future<Output = Result<RangeResponse, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn put<'life0, 'async_trait>( &'life0 self, req: PutRequest ) -> Pin<Box<dyn Future<Output = Result<PutResponse, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn batch_put<'life0, 'async_trait>( &'life0 self, req: BatchPutRequest ) -> Pin<Box<dyn Future<Output = Result<BatchPutResponse, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn batch_get<'life0, 'async_trait>( &'life0 self, req: BatchGetRequest ) -> Pin<Box<dyn Future<Output = Result<BatchGetResponse, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn compare_and_put<'life0, 'async_trait>( &'life0 self, req: CompareAndPutRequest ) -> Pin<Box<dyn Future<Output = Result<CompareAndPutResponse, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn delete_range<'life0, 'async_trait>( &'life0 self, req: DeleteRangeRequest ) -> Pin<Box<dyn Future<Output = Result<DeleteRangeResponse, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn batch_delete<'life0, 'async_trait>( &'life0 self, req: BatchDeleteRequest ) -> Pin<Box<dyn Future<Output = Result<BatchDeleteResponse, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn move_value<'life0, 'async_trait>( &'life0 self, req: MoveValueRequest ) -> Pin<Box<dyn Future<Output = Result<MoveValueResponse, Self::Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

MoveValue atomically renames the key to the given updated key.

Provided Methods

fn get<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = Result<Option<KeyValue>, Self::Error>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn exists<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Check if the key exists, not returning the value. If the value is large, this method is more efficient than get.

fn delete<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 [u8], prev_kv: bool ) -> Pin<Box<dyn Future<Output = Result<Option<KeyValue>, Self::Error>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,