common_meta

Module peer

Structs

NoopPeerAllocator

Peer

Traits

PeerAllocator

PeerAllocator allocates Peers for creating region or flow.

PeerDiscovery

PeerDiscovery is a service for discovering active peers in the cluster.

PeerResolver

PeerResolver provides methods to look up peer information by node ID.

Type Aliases

PeerAllocatorRef

PeerDiscoveryRef

PeerResolverRef