common_meta::reconciliation

Module utils

Structs

Context

LogicalTableMeta

Metadata for an inflight logical table subprocedure.

PartialRegionMetadata 🔒

PartialSuccessResult

The result of waiting for inflight subprocedures.

PhysicalTableMeta

Metadata for an inflight physical table subprocedure.

ReconcileCatalogMetrics

The metrics of reconciling catalog.

ReconcileDatabaseMeta

Metadata for an inflight database subprocedure.

ReconcileDatabaseMetrics

The metrics of reconciling database.

ReconcileLogicalTableMetrics

The metrics of reconciling logical tables.

ReconcileTableMetrics

The metrics of reconciling physical tables.

Enums

ResolveColumnMetadataResult

The result of resolving column metadata.

SubprocedureMeta

The inflight subprocedure metadata.

WaitForInflightSubproceduresResult

The result of waiting for inflight subprocedures.

Constants

COLUMN_METADATA_CONSISTENT 🔒

COLUMN_METADATA_INCONSISTENT 🔒

CREATE_TABLES 🔒

UPDATE_TABLE_INFO 🔒

Functions

build_column_metadata_from_table_info 🔒

Constructs a vector of [ColumnMetadata] from the provided table information.

build_table_meta_from_column_metadatas 🔒

Builds a [RawTableMeta] from the provided [ColumnMetadata]s.

check_column_metadata_invariants 🔒

Checks whether the schema invariants hold between the existing and new column metadata.

check_column_metadatas_consistent 🔒

Checks if the column metadatas are consistent.

need_update_logical_table_info 🔒

Returns true if the logical table info needs to be updated.

resolve_column_metadatas_with_latest 🔒

Resolves column metadata inconsistencies among the given region metadatas by selecting the column metadata with the highest schema version.

resolve_column_metadatas_with_metasrv 🔒

Resolves column metadata inconsistencies among the given region metadatas by using the column metadata from the metasrv as the source of truth.

wait_for_inflight_subprocedures 🔒

Wait for inflight subprocedures.