common_meta::reconciliation::utils

Function need_update_logical_table_info

pub(crate) fn need_update_logical_table_info(
    table_info: &RawTableInfo,
    column_metadatas: &[ColumnMetadata],
) -> bool

Returns true if the logical table info needs to be updated.

The logical table only support to add columns, so we can check the length of column metadatas to determine whether the logical table info needs to be updated.