common_meta

Module region_registry

Structs

LeaderRegion

Represents information about a leader region in the cluster. Contains the datanode id where the leader is located, and the current manifest version.

LeaderRegionRegistry

Registry that maintains a mapping of all leader regions in the cluster. Tracks which datanode is hosting the leader for each region and the corresponding manifest version.

Enums

LeaderRegionManifestInfo

Type Aliases

LeaderRegionRegistryRef