common_meta::rpc

Module ddl

Structs

AlterDatabaseTask

AlterTableTask

CommentOnTask

Comment on table, column, or flow

CreateDatabaseTask

CreateFlowTask

Create flow

CreateTableTask

CreateViewTask

A CREATE VIEW task.

DropDatabaseTask

DropFlowTask

Drop flow

DropTableTask

DropViewTask

A DROP VIEW task.

FlowQueryContext

Lightweight query context for flow operations containing only essential fields. This is a subset of QueryContext that includes only the fields actually needed for flow creation and execution.

QueryContext

SetDatabaseOptions

SubmitDdlTaskRequest

SubmitDdlTaskResponse

TruncateTableTask

UnsetDatabaseOptions

Enums

AlterDatabaseKind

CommentObjectId

Represents the ID of the object being commented on (Table or Flow).

CommentObjectType

DdlTask

DDL tasks

SetDatabaseOption

UnsetDatabaseOption

Constants

TTL_KEY 🔒