common_meta::rpc

Module router

Structs

Partition

Region

RegionRoute

RegionRouteBuilder

Builder for RegionRoute.

RegionRoutes

Table

TableRoute

Enums

LeaderState

The State of the Region Leader. TODO(dennis): It’s better to add more fine-grained statuses such as PENDING etc.

RegionRouteBuilderError

Error type for RegionRouteBuilder

Functions

as_utf8 🔒

as_utf8_vec 🔒

convert_to_region_leader_map

Returns the HashMap<[RegionNumber], &Peer>;

find_follower_regions

Returns the region numbers of the follower regions on the target datanode.

find_followers

Returns the followers of the table.

find_leader_regions

Returns the region numbers of the leader regions on the target datanode.

find_leaders

Returns the leader peers of the table.

find_region_leader

from_utf8

from_utf8_vec

operating_leader_regions

Returns the operating leader regions with corresponding DatanodeId.

region_distribution

Returns the distribution of regions to datanodes.