common_meta::rpc::router

Function region_distribution

pub fn region_distribution(region_routes: &[RegionRoute]) -> RegionDistribution

Returns the distribution of regions to datanodes.

The distribution is a map of datanode id to a list of region ids. The list of region ids is sorted in ascending order.