common_meta

Module snapshot

Modules

file

Structs

FileExtension

FileName

MetadataSnapshotManager

The manager of the metadata snapshot.

Enums

DataType

FileFormat

The format of the backup file.

Constants

MAX_REQUEST_SIZE 🔒

The maximum size of the request to put metadata, use 1MiB by default.

Functions

is_internal_key 🔒

Returns true if the key is an internal key.