common_meta

Module datanode

Structs

DATANODE_LEASE_KEY_PATTERN 🔒

DATANODE_STAT_KEY_PATTERN 🔒

DatanodeStatKey

The key of the datanode stat in the memory store.

DatanodeStatValue

The value of the datanode stat in the memory store.

GcStat

INACTIVE_REGION_KEY_PATTERN 🔒

RegionStat

The statistics of a region.

Stat

The key of the datanode stat in the storage.

TopicStat

Enums

RegionManifestInfo

Constants

DATANODE_STAT_PREFIX 🔒

REGION_STATISTIC_KEY

Traits

TopicStatsReporter

Trait for reporting statistics about topics.