common_meta::ddl::alter_table::region_request

Function create_proto_alter_kind

fn create_proto_alter_kind(
    table_info: &RawTableInfo,
    alter_kind: &Kind,
) -> Result<Option<Kind>>

Creates region proto alter kind from table_info and alter_kind.

It always adds column if not exists and drops column if exists. It skips the column if it already exists in the table.