common_meta

Module distributed_time_constants

Constants

DATANODE_LEASE_SECS

When creating table or region failover, a target node needs to be selected. If the node’s lease has expired, the Selector will not select it.

FLOWNODE_LEASE_SECS

FRONTEND_HEARTBEAT_INTERVAL_MILLIS

The frontend will also send heartbeats to Metasrv, sending an empty heartbeat every HEARTBEAT_INTERVAL_MILLIS * 6 seconds.

HEARTBEAT_INTERVAL_MILLIS

Heartbeat interval time (is the basic unit of various time).

MAILBOX_RTT_SECS

The default mailbox round-trip timeout.

META_KEEP_ALIVE_INTERVAL_SECS

In a lease, there are two opportunities for renewal.

META_LEASE_SECS

The lease seconds of metasrv leader.

POSTGRES_KEEP_ALIVE_SECS

The keep-alive interval of the Postgres connection.

REGION_LEASE_SECS

The lease seconds of a region. It’s set by 3 heartbeat intervals (HEARTBEAT_INTERVAL_MILLIS × 3), plus some extra buffer (1 second).