common_meta::key::flow

Module flow_state

Structs

FlowStat

Flow’s state report, send regularly through heartbeat message

FlowStateKey

The key stores the state size of the flow.

FlowStateKeyInner 🔒

The key of flow state.

FlowStateManager

The manager of FlowStateKey. Since state size changes frequently, we store it in memory.

FlowStateValue

The value of flow state size

Constants

FLOW_STATE_KEY 🔒

The entire FlowId to Flow Size’s Map is stored directly in the value part of the key.

Type Aliases

FlowStateManagerRef