Module common_meta::lock_key

Enums

CatalogLock

CatalogLock acquires the lock on the tenant level.

FlowLock

FlowLock acquires the lock on the table level.

FlowNameLock

FlowNameLock prevents any procedures trying to create a flow named it.

RegionLock

RegionLock acquires the lock on the region level.

SchemaLock

SchemaLock acquires the lock on the database level.

TableLock

TableLock acquires the lock on the table level.

TableNameLock

TableNameLock prevents any procedures trying to create a table named it.

Constants

CATALOG_LOCK_PREFIX 🔒

FLOW_LOCK_PREFIX 🔒

FLOW_NAME_LOCK_PREFIX 🔒

REGION_LOCK_PREFIX 🔒

SCHEMA_LOCK_PREFIX 🔒

TABLE_LOCK_PREFIX 🔒

TABLE_NAME_LOCK_PREFIX 🔒