common_meta::reconciliation::utils

Function wait_for_inflight_subprocedures

pub(crate) async fn wait_for_inflight_subprocedures<'a>(
    procedure_ctx: &Context,
    subprocedures: &'a [SubprocedureMeta],
    fail_fast: bool,
) -> Result<WaitForInflightSubproceduresResult<'a>>

Wait for inflight subprocedures.

If fail_fast is true, the function will return an error if any subprocedure fails. Otherwise, the function will continue waiting for all subprocedures to complete.